requirements.txt
python-telegram-bot==20.8
aiohttp==3.9.5
requests==2.31.0
aiosqlite==0.20.0
python-dotenv==1.0.1
